# Conflicts:
#	src/org/mozilla/javascript/NativeJavaList.java
#	src/org/mozilla/javascript/NativeJavaMap.java

+++ src/org/mozilla/javascript/NativeJavaMap.java
@@ -14,6 +14,17 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.function.IntFunction;
 
+/**
+ * <code>NativeJavaMap</code> is a wrapper for java objects implementing <code>java.util.Map
+ * </code> interface. When {@link Context#FEATURE_ENABLE_JAVA_MAP_ACCESS} is enabled, property based
+ * access like <code>map[key]</code> is delegated to {@link Map#get(Object)} or {@link
+ * Map#put(Object, Object)} operations so that a <code>JavaMap</code> acts very similar to a
+ * javascript <code>Object</code> There is also an iterator to iterate over entries with <code>
+ * for .. of</code>.
+ *
+ * <p><b>Limitations:</b> The wrapped map should have <code>String</code> or <code>Integer</code> as
+ * key. Otherwise, property based access may not work properly.
+ */
 public class NativeJavaMap extends NativeJavaObject {
 
     private static final long serialVersionUID = 46513864372878618L;
@@ -145,7 +156,7 @@ public class NativeJavaMap extends NativeJavaObject {
             List<Object> ids = new ArrayList<>(map.size());
             for (Object key : map.keySet()) {
                 if (key instanceof Integer) {
-                    ids.add((Integer) key);
+                    ids.add(key);
                 } else {
                     ids.add(ScriptRuntime.toString(key));
                 }
